+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 1. [](#bugfix)
     * Fixed user login / remember me triggering before admin gets initialized
+    * Fixed a bug when deleting files via AJAX
 
 # v1.9.0-beta.6
 ## 11/12/2018